requirements.txt
requests  
numpy
pyaudio~=0.2.11
websockets~=10.4
ws4py~=0.5.1
PyQt5==5.15.10
PyQt5-sip==12.13.0
PyQtWebEngine==5.15.6 
flask~=3.0.0
openpyxl~=3.0.9
flask_cors~=3.0.10
websocket-client
azure-cognitiveservices-speech
aliyun-python-sdk-core
simhash
pytz
gevent
edge_tts
pydub
tenacity==8.2.3
pygame
scipy
flask-httpauth
opencv-python
psutil
langchain
langchain_openai
langgraph
bs4
schedule
mcp
docx